2025/7 part 2 works, overflow issue fixed

+++ src/year2025/day7/GridElement.java
@@ -3,8 +3,8 @@ package year2025.day7;
 class GridElement {
     int x;
     int y;
-    int overlapCounter;
+    long overlapCounter;
     char type;
     int noMoreCollision; // must be 1 or 2
-    int incomingBeamN;
+    long incomingBeamN;
 }
